feat: migrate to Next.js with GitHub OAuth and n8n integration
Major architecture upgrade from Vite+React to Next.js 15:

✨ Features:
- Next.js 15 with App Router and TypeScript
- GitHub OAuth authentication via Auth.js/NextAuth v5
- Whitelist-based access control (GitHub usernames + email domains)
- Server-side n8n webhook integration (/api/n8n/trigger)
- Route protection middleware (unauthorized users -> /login)
- Preserved all 4 existing apps with full functionality
- Bilingual documentation (README.md + README.zh-CN.md)

🔒 Authentication:
- GitHub OAuth Provider with Auth.js
- Environment-based whitelist (ALLOWED_GITHUB_LOGINS, ALLOWED_EMAIL_DOMAINS)
- Automatic redirect to /login for unauthenticated users
- /unauthorized page for non-whitelisted users
- Secure session management

🚀 Architecture:
- Migrated from Vite SPA to Next.js SSR/SSG
- App Router with proper layouts and nested routes
- Server-side API routes for secure n8n integration
- Middleware for route protection
- Tailwind CSS for styling
- TypeScript strict mode

📦 Apps Migrated:
- keyword-dedup: Keyword deduplication tool
- csv-converter: CSV/TSV converter with table preview
- profit-calculator: Amazon FBA profit calculator
- asin-keywords: ASIN keyword lookup (now uses server-side API)

🔧 Configuration:
- .env.example with all required variables
- next.config.ts with standalone output
- tailwind.config.ts for styling
- Updated .gitignore for Next.js
- tsconfig.json for Next.js + TypeScript

📚 Documentation:
- README.md: English version with deployment guide
- README.zh-CN.md: Detailed Chinese version
- Environment setup instructions
- GitHub OAuth app creation guide
- Vercel deployment instructions
- n8n webhook integration guide

🗑️ Cleanup:
- Removed old Vite configuration (vite.config.ts, tsconfig.node.json)
- Removed index.html (Next.js doesn't need it)
- Removed src/ directory (migrated to app/ + apps/)
- Updated package.json with Next.js scripts

🔗 Integration:
- Server-side n8n API endpoint with authentication
- Secure webhook forwarding (secrets in env vars)
- ASIN keywords app updated to use /api/n8n/trigger
- No client-side exposure of sensitive URLs

Closes #9
