+++ .env.example
@@ -1,0 +1,18 @@
+# Auth.js Configuration
+# Generate a secret with: openssl rand -base64 32
+AUTH_SECRET=your-auth-secret-here
+
+# GitHub OAuth (Create at: https://github.com/settings/developers)
+AUTH_GITHUB_ID=your-github-oauth-app-client-id
+AUTH_GITHUB_SECRET=your-github-oauth-app-client-secret
+
+# Whitelist Configuration (at least one required)
+# Option 1: GitHub usernames (comma-separated, case-insensitive)
+ALLOWED_GITHUB_LOGINS=username1,username2,username3
+
+# Option 2: Email domains (comma-separated, case-insensitive)
+ALLOWED_EMAIL_DOMAINS=example.com,company.com
+
+# n8n Webhook Configuration
+N8N_WEBHOOK_URL=https://your-n8n-instance.com/webhook/your-webhook-id
+N8N_WEBHOOK_SECRET=your-optional-webhook-secret
