+++ apps/asin-keywords/meta.ts
@@ -1,0 +1,10 @@
+import type { AppMeta } from '@/lib/types';
+
+export const meta: AppMeta = {
+  slug: 'asin-keywords',
+  name: 'ASIN → 关键词（DataForSEO）',
+  description: '输入 Amazon ASIN，通过 DataForSEO 获取该产品的排名关键词列表',
+  category: 'Keyword Research',
+  tags: ['ASIN', 'DataForSEO', '关键词排名', 'Amazon'],
+  order: 2,
+};
