+++ apps/csv-converter/meta.ts
@@ -1,0 +1,10 @@
+import type { AppMeta } from '@/lib/types';
+
+export const meta: AppMeta = {
+  slug: 'csv-converter',
+  name: 'CSV/TSV转换器',
+  description: '输入CSV或TSV文本，预览表格数据，导出为CSV',
+  category: 'Data Tools',
+  tags: ['CSV', 'TSV', '数据', '转换', '导出'],
+  order: 1,
+};
