+++ apps/keyword-dedup/meta.ts
@@ -1,0 +1,10 @@
+import type { AppMeta } from '@/lib/types';
+
+export const meta: AppMeta = {
+  slug: 'keyword-dedup',
+  name: '关键词去重工具',
+  description: '输入多行关键词,自动去重、排序,一键复制结果',
+  category: 'Keyword Research',
+  tags: ['关键词', '去重', '排序', '工具'],
+  order: 1,
+};
