+++ next.config.ts
@@ -1,0 +1,8 @@
+import type { NextConfig } from 'next';
+
+const nextConfig: NextConfig = {
+  output: 'standalone',
+  experimental: {},
+};
+
+export default nextConfig;
