--- src/apps/profit-calculator/meta.ts
@@ -1,10 +1,0 @@
-import type { AppMeta } from '../../types';
-
-export const meta: AppMeta = {
-  slug: 'profit-calculator',
-  name: '利润计算器',
-  description: '输入售价、成本、平台费率和运费，快速计算利润与毛利率',
-  category: 'Profit & FBA',
-  tags: ['利润', '计算器', 'FBA', '成本'],
-  order: 1,
-};
